+++ src/Storage/EverTask.Storage.Postgres/PostgresTaskStoreContext.cs
@@ -1,0 +1,5 @@
+namespace EverTask.Storage.Postgres;
+
+public class PostgresTaskStoreContext(
+    DbContextOptions<PostgresTaskStoreContext> options,
+    IOptions<ITaskStoreOptions> storeOptions) : TaskStoreEfDbContext<PostgresTaskStoreContext>(options, storeOptions);